Move share execution effects into useShareExecution
ShareSection.tsx still owned all of the channel-side effects: the
auto-skip dispatch, the START_SHARE emit, the share-auth-completed
telemetry, the progress reducer plumbing, and the cancel/publish
handlers. The progress effect also needed an eslint-disable on its dep
list because it intentionally only re-ran on shareProgress changes.

Extract the lot into a useShareExecution hook. Latest reducer/git state
flows in through a ref, so the progress effect can drop the
eslint-disable while still firing only on shareProgress changes.
ShareSection now reads as plain view routing + the hook contract.
